feat: v1.5 — warm reskin, WhisperKit 1.0, reliability fixes
UI: full reskin to match openwhisperer.com — warm cream/gold dynamic
light/dark palette (Theme.swift), bundled Fraunces SemiBold wordmark,
warm brand status colors (no clashing system green/cyan/orange),
gold-tinted card headers, refined cards/pickers/buttons; transcription
overlay rethemed; removed redundant Setup diagnostic rows; new menu
screenshot.

Engines: WhisperKit 0.18.0 -> 1.0.0 (FluidAudio already latest).
Cache-aware first-run STT messaging (download vs ANE compile) so a slow
first launch no longer looks stuck.

Reliability (from architect + swift review): TTS generation-guard +
playback-error lock clear; loopback TTS server body-size cap + surfaced
bind-failure; KokoroTTS voice-pack error logging; AVAudioFormat guard;
uniform 900s voice-turn freshness across hooks.

Docs: README reframed to v1.5 + Hakan Ensari acknowledgment for the
native rewrite; AGENTS tagless limitations; distribution/notarization
runbook; UX backlog; v1.5 plan. Version 1.4.0 -> 1.5.0.

+++ app/Tests/OpenWhispererKitTests/PCMConversionChecks.swift
@@ -18,7 +18,7 @@ func pcmConversionFailures() -> [String] {
     expect([-32768], [-1.0], "minClipsToNegativeOne")
     expect([16384], [0.5], "half")
     expect([-16384], [-0.5], "negativeHalf")
-    expect([32767], [32767.0 / 32768.0], "maxNearPositiveOne")
+    expect([32767], [Float(32767.0 / 32768.0)], "maxNearPositiveOne")
     expect([0, 16384, -16384, -32768], [0, 0.5, -0.5, -1.0], "sequence")
 
     return failures
